Implement usage of middlewares on new rack app
